refactor(src): TemplateListItem 타입 types index에서 export하도록 변경
