+++ frontend/src/components/TemplateCard/TemplateCard.tsx
@@ -5,8 +5,7 @@ import { oneLight } from 'react-syntax-highlighter/dist/esm/styles/prism';
 import { Button, Flex, TagButton, Text } from '@/components';
 import { useToggle } from '@/hooks/utils';
 import { theme } from '@/style/theme';
-import type { Tag } from '@/types';
-import { TemplateListItem } from '@/types/template';
+import type { Tag, TemplateListItem } from '@/types';
 import { getLanguageByFilename } from '@/utils';
 import { formatRelativeTime } from '@/utils/formatRelativeTime';
 import * as S from './TemplateCard.style';
